fix(docs) tricks to correct sphinx warning - some remain, but the autoapi is at least now generating all pages
