+++ src/plaid/types/feature_types.py
@@ -14,12 +14,12 @@ try:
 except ImportError:  # pragma: no cover
     from typing_extensions import TypeAlias
 
-from plaid.types import Array
+from plaid.types.common import Array
 
 # Physical data types
 ScalarType: TypeAlias = Union[float, int]
 FieldType: TypeAlias = Array
-TimeSequenceType: TypeAlias = Array
+TimeSequenceType: TypeAlias = FieldType
 TimeSeriesType: TypeAlias = Tuple[TimeSequenceType, FieldType]
 
 # Feature data types
